# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenAgriNet API Specification
  version: 1.0.0
  description: |-
    Three surfaces, one network. **Discovery** publishes resources and searches them; the
    **registry** says who is on the network and how to call them; the **provider node**
    returns the data itself.

    They are separate services on separate hosts. Every path carries its own
    `servers`, so the host is unambiguous per operation.

    | surface | host | auth | tags |
    |---|---|---|---|
    | Discovery | `discovery-network-vistaar.da.gov.in` | handled by the adapter in front | Publish, Discover |
    | Registry | `registry-network-vistaar.da.gov.in` | bearer | SchemaRegistry, Participant, ProviderSchema |
    | Provider node | `provider-network-vistaar.da.gov.in` | Beckn request signature, verified by the adapter | Select |

    Discovery follows Beckn v2.0.0; see `beckn.yaml` for the protocol and for the
    authoritative shape of `Catalog`, `Provider`, `Resource` and `Offer`, which are
    declared permissively here. Outcomes are returned as the 200 body.

    The registry is three tables:

    | table | what it holds |
    |---|---|
    | `SchemaRegistry` | The kinds of data the network knows about — OpenAgriNet's network-spec metadata. One row per capability: what a weather observation or a mandi price is, and where its schema lives. |
...

    Its REST surface is generated by Sunbird RC from the JSON Schema files
...
servers:
  - &id001
...
  - &id002
...
  - &id003
...
paths:
  /publish:
    post:
      tags:
        - Publish
      operationId: publish
      summary: Publish or update resources
      description: |-
        Submits resources — the things a farmer can be shown. On the wire they travel grouped
        into one or more catalogs: a catalog is the unit of submission and of verdict, and the
        resources inside it are what `/discover` later matches on.

        Each catalog is processed independently and gets its own verdict — `ACCEPTED`, `REJECTED`
        or `PARTIAL` — in the 200 body. A `PARTIAL` means it landed with some of its resources
        dropped; the `errors` array says which and why.

        Phase 1 accepts regular resources only. A directive naming `catalogType: MASTER`, or a
        resource carrying `extends`, is rejected with `SCH_TYPE_NOT_SUPPORTED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
            examples:
              weatherForecast:
                summary: One weather resource, MERGE
                value:
                  context:
                    action: publish
                    version: 2.0.0
                    messageId: 1c0a55d7-8e64-4b19-9a2f-33b7c6e1d905
                    transactionId: 9f2c1a8e-4b70-4d31-9c55-6f2e0b1d7a44
                    timestamp: "2026-08-26T06:11:58.004Z"
                  message:
                    catalogs:
                      - id: cat:mausamgram:weather
                        isActive: true
                        provider:
                          id: mausamgram
                          descriptor:
                            code: IMD-NWP-01
                            name: IMD Mausamgram NWP
                          availableAt:
                            - geo:
                                type: Polygon
                                coordinates:
                                  - - - 68.1
                                      - 8.0
                                    - - 97.4
                                      - 8.0
                                    - - 97.4
                                      - 37.1
                                    - - 68.1
                                      - 37.1
                                    - - 68.1
                                      - 8.0
                        validity:
                          startDate: "2026-01-01T00:00:00Z"
                          endDate: "2026-03-31T23:59:59Z"
                        resources:
                          - id: res:mausamgram:point-forecast
                            descriptor:
                              name: Five-day point forecast
                            resourceAttributes:
                              "@context": https://schemas.openagrinet.global/schema/WeatherObservation/v0.1/context.jsonld
                              "@type": openagrinet:WeatherObservation
                              informationMode: OnDemand
                              supportedObservationTypes:
                                - Forecast
                              supportedParameters:
                                - Rainfall
                                - Temperature
                                - Humidity
                                - WindSpeed
                                - WindDirection
                              geographicGranularity:
                                - Point
                              forecastHorizon: P5D
                              updateFrequency: PT12H
                              subjectCategories:
                                - Weather
                    publishDirectives:
                      - catalogId: cat:mausamgram:weather
                        catalogType: REGULAR
                        updateMode: MERGE
      responses:
        "200":
          description: |
            ACK. Per-catalog verdicts in the body. A 200 does **not** mean every catalog
            was accepted — read `message.results[].status`.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-Response-Time:
              $ref: "#/components/headers/XResponseTime"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnPublishResponse"
              examples:
                mixed:
                  summary: One accepted, one rejected, one partial
                  value:
                    context:
                      action: catalog/on_publish
                      version: 2.0.0
                      messageId: 1c0a55d7-8e64-4b19-9a2f-33b7c6e1d905
                      transactionId: 9f2c1a8e-4b70-4d31-9c55-6f2e0b1d7a44
                      timestamp: "2026-08-26T06:11:58.612Z"
                    message:
                      results:
                        - catalogId: cat:mausamgram:weather
                          status: ACCEPTED
                          stats:
                            itemCount: 1
                            providerCount: 1
                            categoryCount: 1
                        - catalogId: cat-2
                          status: REJECTED
                          errors:
                            - code: SCH_TYPE_NOT_SUPPORTED
                              message: master catalogs are out of scope in phase 1
                              details:
                                path: $.message.publishDirectives[1]
                        - catalogId: cat-3
                          status: PARTIAL
                          stats:
                            itemCount: 8
                            providerCount: 1
                            categoryCount: 1
                          errors:
                            - code: SCH_INVALID_FORMAT
                              message: geometry could not be read
                              details:
                                path: $.message.catalogs[2]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    servers:
      - *id001
  /discover:
    post:
      tags:
        - Discover
      operationId: discover
      summary: Search the network for resources
      description: |-
        Matches published resources against an intent and returns those that match, in the
        catalogs that carry them, in the 200 body. Answered from the local index; this route
        never calls providers or upstream APIs.

        `context.networkId` applies no network predicate here — every network's resources are
        eligible. On `/publish` the same field fills an empty `visibleTo`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverRequest"
            examples:
              rainNearNashik:
                summary: Weather observations within 250 km of Nashik
                value:
                  context:
                    action: discover
                    version: 2.0.0
                    messageId: 1c0a55d7-8e64-4b19-9a2f-33b7c6e1d905
                    transactionId: 9f2c1a8e-4b70-4d31-9c55-6f2e0b1d7a44
                    timestamp: "2026-08-26T06:11:58.004Z"
                    schemaContext:
                      - https://schemas.openagrinet.global/schema/WeatherObservation/v0.1/context.jsonld#openagrinet:WeatherObservation
                  message:
                    intent:
                      textSearch: weather forecast rain next five days
                      filters:
                        type: jsonpath
                        expression: $.catalogs[*].resources[*] ? (@.resourceAttributes."@type" == "openagrinet:WeatherObservation")
                      spatial:
                        - op: S_DWITHIN
                          targets: $.catalogs[*].provider.availableAt[*].geo
                          geometry:
                            type: Point
                            coordinates:
                              - 73.7898
                              - 19.9975
                          distanceMeters: 250000
                          quantifier: ANY
                          srid: EPSG:4326
      responses:
        "200":
          description: |-
            ACK. The matching resources, in the `catalogs` array that carries them. An empty array
            is an honest "nothing matched", not a failure.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-Response-Time:
              $ref: "#/components/headers/XResponseTime"
            X-Beckn-Degraded:
              $ref: "#/components/headers/XBecknDegraded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnDiscoverResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    servers:
      - *id001
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        Bearer token. `_osConfig.roles` gates the entity rather than the verb, so a token that
        can read these records can also write them.
...
  headers:
    XRequestId:
      description: |
        Echoed from the request's `X-Request-Id` when the caller sent one, otherwise
        minted here. Honouring the caller's value is what lets one caller correlate a
        request across their logs and ours.
      schema:
        type: string
    XResponseTime:
      description: Server-side handling time.
      schema:
        type: string
        example: 12ms
    XBecknErrorType:
      description: |-
        The fault category — `context`, `core`, `domain`, `policy`, `system`. Travels as a header;
        `ERROR_INCLUDE_LEGACY_TYPE=true` also writes it into the body as `error.type` for
        v1-style clients. Defaults to false.
      schema:
        type: string
        enum:
          - context
          - core
          - domain
          - policy
          - system
    XBecknDegraded:
      description: |-
        Comma-separated retrieval modes that were unavailable for this request, when the answer
        was served without them.
      schema:
        type: string
        example: semantic
    RetryAfter:
      description: Seconds to wait before retrying, computed by the limiter.
      schema:
        type: integer
        example: 3
  responses:
    BadRequest:
      description: |-
        NACK. The request is malformed or refused on content: every `CTX_` and `SCH_` code, the
        `BIZ_` family, and `NET_CATALOG_SOURCE_UNAVAILABLE`. Retrying does not help.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
          examples:
            badJsonPath:
              summary: An RFC 9535 filter expression, which is not executed
              value:
                message:
                  status: NACK
                  messageId: 1c0a55d7-8e64-4b19-9a2f-33b7c6e1d905
                  error:
                    code: SCH_INVALID_JSONPATH
                    message: RFC 9535 syntax is not executed; use PostgreSQL SQL/JSON path with a ? predicate
                    details:
                      path: $.message.intent.filters.expression
    Unauthorized:
      description: |-
        NACK. The `AUT_` family, minus the rate limiter. Signature verification runs in the
        adapter in front of this service, not here.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
    Forbidden:
      description: NACK. The `POL_` family, except the oversized-body case which is a 413.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
    PayloadTooLarge:
      description: NACK, `POL_NP_CAPACITY_EXCEEDED` — the body exceeded `SERVER_MAX_REQUEST_BODY_BYTES`.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
    RateLimited:
      description: NACK, `AUT_RATE_LIMITED`. The credentials were fine; the pace was not.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
    ServerError:
      description: NACK, the `NET_` family. The message is generic; the original goes to the log.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-Beckn-Error-Type:
          $ref: "#/components/headers/XBecknErrorType"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Nack"
...
  schemas:
    Context:
      type: object
      description: |-
        The Beckn envelope header. A response echoes `transactionId`, `messageId` and the rest;
        only `action` changes and `timestamp` advances.
      properties:
        action:
          type: string
          description: |-
            Both publish spellings are accepted on `POST /publish` — the schema index is keyed by
            action rather than by URL.
          enum:
            - publish
            - catalog/publish
            - discover
            - catalog/on_publish
            - on_discover
        version:
          type: string
          const: 2.0.0
          description: The one protocol version this build serves.
        senderId:
          type: string
          description: |-
            A DID resolving to the document carrying the sender's verification keys. Not verified at
            this layer.
        receiverId:
          type: string
          description: See `senderId`. Also unverified.
        transactionId:
          type: string
          format: uuid
          description: UUID v4. Constant across every hop of one farmer question.
        messageId:
          type: string
          format: uuid
          description: UUID v4. New per request; echoed on the response and on a NACK.
        timestamp:
          type: string
          format: date-time
          description: |
            RFC 3339, carried as a string rather than a typed time so a bad value can come
            back as a fault naming `$.context.timestamp` instead of failing the whole decode.
        networkId:
          type: string
          description: |-
            The subnet this request belongs to. On `/discover` it applies **no predicate at all** —
            every network's resources match. On `/publish` it fills an empty `visibleTo`.
          example: da.gov.in/vistaar
        schemaContext:
          type: array
          items:
            type: string
            format: uri
          description: |
            JSON-LD context URIs, each with an optional `#fragment` naming the `@type`. A
            **Context** field, never an Intent one — the reference implementation moved it to
            `message.intent`, which `Intent`'s `additionalProperties: false` forbids outright.
        key:
          type: string
        try:
          type: boolean
        ttl:
          type: string
          description: ISO 8601 duration.
        requestDigest:
          description: Opaque. Passed through.
      additionalProperties: true
    PublishRequest:
      type: object
      required:
        - context
        - message
      properties:
        context:
          $ref: "#/components/schemas/Context"
        message:
          $ref: "#/components/schemas/CatalogPublishAction"
    CatalogPublishAction:
      type: object
      required:
        - catalogs
      properties:
        catalogs:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Catalog"
        publishDirectives:
          type: array
          items:
            $ref: "#/components/schemas/PublishDirective"
          description: |-
            Per-catalog processing instructions, matched to a catalog by `catalogId`. Every other
            field has a declared default, so an absent directive and one naming only `catalogId`
            behave identically.
    PublishDirective:
      type: object
      required:
        - catalogId
      properties:
        catalogId:
          type: string
          description: MUST exactly match an `id` in this request's `catalogs` array.
        catalogType:
          type: string
          enum:
            - REGULAR
            - MASTER
          default: REGULAR
          description: |
            **`MASTER` is refused in phase 1** with `SCH_TYPE_NOT_SUPPORTED` — rejected, not
            partially handled.
        updateMode:
          type: string
          enum:
            - MERGE
            - FULL
          default: MERGE
          description: |
            `MERGE` upserts resources by id and retains those this request did not mention.
            `FULL` replaces the catalog. The default matters: `FULL` **deletes** what the
            payload omitted, so an omitted `updateMode` reading as `FULL` would turn every
            directive-less republish into a partial wipe.
        visibleTo:
          type: array
          items:
            type: string
          description: |-
            Restricts this catalog to the listed participant ids. Empty or omitted resolves to the
            request's own `context.networkId`.
        resourceDirectives:
          type: array
          items:
            type: object
          description:
            Links resources to master resources they inherit from. Phase 1 accepts only an
            empty array.
    OnPublishResponse:
      type: object
      required:
        - context
        - message
      properties:
        context:
          $ref: "#/components/schemas/Context"
        message:
          $ref: "#/components/schemas/CatalogOnPublishAction"
    CatalogOnPublishAction:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/CatalogProcessingResult"
    CatalogProcessingResult:
      type: object
      required:
        - catalogId
        - status
      properties:
        catalogId:
          type: string
        status:
          type: string
          enum:
            - ACCEPTED
            - REJECTED
            - PARTIAL
          description:
            "`PARTIAL` means the catalog landed with something dropped; the `errors` array
            says what."
        stats:
          $ref: "#/components/schemas/CatalogStats"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
          description: |
            The spec's own array — which is why the publish path never packs faults into a
            `details.cause` chain. A NACK is the only place the chain is used.
    CatalogStats:
      type: object
      properties:
        itemCount:
          type: integer
        providerCount:
          type: integer
        categoryCount:
          type: integer
    DiscoverRequest:
      type: object
      required:
        - context
        - message
      properties:
        context:
          $ref: "#/components/schemas/Context"
        message:
          type: object
          required:
            - intent
          properties:
            intent:
              $ref: "#/components/schemas/Intent"
    OnDiscoverResponse:
      type: object
      required:
        - context
        - message
      properties:
        context:
          $ref: "#/components/schemas/Context"
        message:
          type: object
          required:
            - catalogs
          additionalProperties: false
          description: |-
            Carries `catalogs` and nothing else — the v2.0.0 schema declares
            `additionalProperties: false`. Degraded modes travel in `X-Beckn-Degraded`.
          properties:
            catalogs:
              type: array
              items:
                $ref: "#/components/schemas/Catalog"
    Intent:
      type: object
      additionalProperties: false
      description: Closed. `schemaContext` belongs on Context, not here.
      properties:
        textSearch:
          type: string
          description: >-
            Free text. Lexical by default; the semantic path is
            configuration-gated. So is the field itself — a deployment running
            `SEARCH_ENABLE_TEXT_SEARCH=false` refuses any non-empty value with
            `SCH_TYPE_NOT_SUPPORTED` rather than answering without it, since on
            a text-only intent dropping the term returns the whole corpus.
        filters:
          $ref: "#/components/schemas/Filters"
        spatial:
          type: array
          items:
            $ref: "#/components/schemas/SpatialConstraint"
        mediaSearch:
          type: object
          description: Passed through. See beckn.yaml.
    Filters:
      type: object
      required:
        - type
        - expression
      properties:
        type:
          type: string
          enum:
            - jsonpath
        expression:
          type: string
          description: |-
            PostgreSQL SQL/JSON path only, and it must carry a `?` predicate. An RFC 9535 expression
            is rejected with `SCH_INVALID_JSONPATH`.
          example: $.catalogs[*].resources[*] ? (@.resourceAttributes.observationType == "Forecast")
...
    Catalog:
      type: object
      required:
        - id
      description: |-
        A provider's resources as one publishable unit — the payload of both directions.
        Declared permissively: the service carries these as opaque JSON and validates them at
        the L1 JSON Schema layer. See beckn.yaml for the authoritative shape.
      properties:
        id:
          type: string
        bppId:
          type: string
        bppUri:
          type: string
          format: uri
        descriptor:
          type: object
        provider:
          type: object
        isActive:
          type: boolean
          default: true
          description: Defaulted by the mapper, so an omitted value is not read as `false`.
        validity:
          $ref: "#/components/schemas/Validity"
        resources:
          type: array
          items:
            type: object
            description: |
              A resource. `resourceAttributes` carries the JSON-LD typed payload — `@context`,
              `@type` and the capability's own fields.

              **A resource carrying `extends` is refused in phase 1** with
              `SCH_TYPE_NOT_SUPPORTED`.
        offers:
          type: array
          items:
            type: object
            description: An offer. `addOns` is refused with `SCH_SCHEMA_VALIDATION_FAILED`.
      additionalProperties: true
    Validity:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        startTime:
          type: string
          description: Time with offset, e.g. 09:00:00+05:30
        endTime:
          type: string
...
    Nack:
      type: object
      required:
        - message
      description: |-
        The rejection body — one shape for 400, 401, 403, 413, 429 and 500. There is no `context`
        key; correlate a NACK by `message.messageId`.
      properties:
        message:
          type: object
          required:
            - status
            - error
          properties:
            status:
              type: string
              const: NACK
            messageId:
              type: string
              description: Echoed from the request.
            error:
              $ref: "#/components/schemas/Error"
    Error:
      type: object
      required:
        - code
        - message
      additionalProperties: false
      description: Closed, and `details` is closed to exactly `path` and `cause`.
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: false
          properties:
            path:
              type: string
              description: JSONPath to the offending field.
              example: $.message.intent.filters.expression
            cause:
              allOf:
                - $ref: "#/components/schemas/Error"
              description: |
                **A NACK carrying several faults is one chain, not a list.** The first fault is
                the `Error`; each remaining fault becomes the `details.cause` of the one before.
        type:
          type: string
          enum:
            - context
            - core
            - domain
            - policy
            - system
          description: |-
            Written only when `ERROR_INCLUDE_LEGACY_TYPE=true`, for v1-style clients that require the
            key in the body. Defaults off; the category otherwise travels as `X-Beckn-Error-Type`.
            When on, it is written at every level of the chain.
...